Complete rework of Tags, add tests for applying Tags, add some example fey Tags
Wanted to move away from the idea of having a parser for Tags that uses human-readable text from a Table. Instead we'll just define them in code with the human-readable part as an attribute for a Tag object (effect_text).

New version of Tag is moved to statblock.py and contains some stuff for handling Tags which conflict in some way (stacks, on_stack(), overwrites, overwritten_by, on_overwrite(), remove()). This might be too complex really, but I wanted something to start with that we can refine later if it's too much or doesn't suit our needs.