Fix action parsing, add logging, add tests for action parsing
This fixes updating action descriptions with math in them. This can come
in two forms. Take this example action:

***Psychic Punch.*** Deal {STR + 2} psychic damage, then
deal 2 (1d6 + 3) bludgeoning damage to all creatures within 30 ft.

1. Non-Wizards of the Coast-style in curly brackets, '{STR + 2}'. This
is just for calculating a one-time value from the template. A single
number is substituted, and the calculation is not shown in the final
output statblock.

2. Wizards of the Coast-style in parentheses. There's a total value in
front, and there's a calculation shown in parentheses which uses the
upper-average dice roll. Here, we just fix the total if it doesn't
match the calculation.

If the Statblock's STR is 3, the final version should look like this:

***Psychic Punch.*** Deal 5 psychic damage, then
deal 7 (1d6 + 3) bludgeoning damage to all creatures within 30 ft.

Also added some logging. Might not end up maintaining it or just delete
it. We'll see how useful it ends up being.