Complete rework of web code
Replaced with Flask app (since I am at least a little bit familiar with it). Had a little bit of trouble following Zhenya's page, so I redid the HTML/CSS as well.

Server is in sort-of working state. You can select a tag table, and it will pull tags from a module in tags/.

Do need an actual WSGI server for production in some point in the future.